requirements.txt
accelerate==1.12.0
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.2.1
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.12.0
anytree==2.13.0
apache-tvm-ffi==0.1.5
asgiref==3.11.0
astor==0.8.1
astroid==3.0.3
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1736871640129/work
bce-python-sdk==0.9.55
beautifulsoup4==4.12.3
blake3==1.0.8
boto3==1.34.162
botocore==1.34.162
cachetools==6.2.2
camel-converter==5.0.0
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
comm==0.2.3
compressed-tensors==0.9.3
cryptography==43.0.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
curl_cffi==0.7.4
dashscope==1.19.3
datasets==3.0.0
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
deprecation==2.1.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.2.9
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastapi==0.123.5
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.20.0
fire==0.4.0
flashinfer-python==0.5.3
frozenlist==1.8.0
fsspec==2024.3.1
future==1.0.0
gguf==0.17.1
gitdb==4.0.12
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.4.1
googleapis-common-protos==1.72.0
GPUtil==1.4.0
greenlet==3.2.4
grep-ast==0.3.3
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
gymnasium==0.29.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
imap-tools==1.5.0
importlib_metadata==8.0.0
interegular==0.3.3
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
lancedb==0.4.0
lark==1.2.2
lazy-object-proxy==1.12.0
libcst==1.0.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.6.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
meilisearch==0.21.0
mistral_common==1.8.6
more-itertools==10.8.0
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-cutlass-dsl==4.3.3
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.6.1
openai==1.64.0
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.1.1
parse==1.20.2
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pathable==0.4.4
pathspec==0.12.1
peft==0.12.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
portalocker==2.10.1
prance==23.6.21.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyairports==0.0.1
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1750615949581/work
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-extra-types==2.10.6
pydantic_core==2.23.4
pyee==13.0.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.8.0
pylance==0.9.0
pylint==3.0.4
pymongo==4.15.5
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.1.0
qdrant-client==1.7.0
qianfan==0.4.12.3
qutip==5.2.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
ray==2.52.1
redis==5.0.8
referencing==0.30.2
regex==2023.12.25
requests==2.32.5
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
s3transfer==0.10.4
safetensors==0.7.0
scikit-learn==1.3.2
scipy==1.15.3
semantic-kernel==0.4.3.dev0
semver==3.0.4
sentencepiece==0.2.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
spark-ai-python==0.3.31
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
ta==0.10.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.22.1
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.57.3
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
tree-sitter-python==0.23.6
triton==3.2.0
trl==0.17.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.11.0
tzdata==2025.2
unidiff==0.7.5
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.8.5
volcengine-python-sdk==1.0.130
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==1.15.0
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.6.0
yarl==1.22.0
zhipuai==2.1.5.20250825
zipp==3.23.0
